Rename frenet_planner_node to sampler_node
