+++ planning/sampler_node/launch/sampler_node.xml
@@ -2,9 +2,9 @@
   <arg name="input_objects" default="/perception/object_recognition/objects" />
   <arg name="input_path_topic" default="/planning/scenario_planning/lane_driving/behavior_planning/path" />
   <arg name="output_path_topic" default="output/path" />
-  <arg name="param_path" default="$(find-pkg-share frenet_planner_node)/config/frenet_planner_node.param.yaml" />
+  <arg name="param_path" default="$(find-pkg-share sampler_node)/config/sampler_node.param.yaml" />
 
-  <node pkg="frenet_planner_node" exec="frenet_planner_node_exe" name="frenet_planner_node" output="screen" launch-prefix="konsole -e gdb -ex run --args">
+  <node pkg="sampler_node" exec="sampler_node_exe" name="sampler_node" output="screen" launch-prefix="konsole -e gdb -ex run --args">
     <remap from="~/input/objects" to="$(var input_objects)"/>
     <remap from="~/input/path" to="$(var input_path_topic)" />
     <remap from="~/output/path" to="$(var output_path_topic)" />
